refactor(chat): split renderer runtime state
